# [Mono.Android] remove System.Reflection usage in UncaughtExceptionHandler