--- a/src/Mono.Android/Java.Lang/Object.cs
+++ b/src/Mono.Android/Java.Lang/Object.cs
@@ -7,6 +7,8 @@ using Java.Interop;
 
 using Android.Runtime;
 
+using LogLevel = Android.Runtime.LogLevel;
+
 namespace Java.Lang {
 
 	[DataContract]