--- a/src/Mono.Android/Xamarin.Android.Net/AuthModuleDigest.cs
+++ b/src/Mono.Android/Xamarin.Android.Net/AuthModuleDigest.cs
@@ -7,6 +7,8 @@ using Android.Runtime;
 
 using Java.Net;
 
+using LogLevel = Android.Runtime.LogLevel;
+
 namespace Xamarin.Android.Net
 {
 	sealed class AuthModuleDigest : IAndroidAuthenticationModule