@jonpryor 2019-12-06
> If we lived in a world where `mscorlib.dll` has `InternalsVisibleTo` for `Mono.Android`, 

We *may* be able to live in such a world in the short-term (d16-6) timeframe.

I'm dubious about us *continuing* to live in such a world in the .NET 5 timeframe.  It's entirely plausible that in .NET 5 "everybody" will be using the *same* assemblies, and those assemblies *might not* have `InternalsVisibleTo`.

It's thus potentially dubious.

Furthermore, the Reflection usage only occurs when there's an *actual* unhandled exception, when `UncaughtExceptionHandler.UncaughtException()` is executed.  We're speeding up what *should* be a rare occurrence (probably followed quickly by process death).