@jonathanpeppers 2019-12-06
> We're speeding up what should be a rare occurrence (probably followed quickly by process death).

I am now wondering how I saw a speedup. I ran a script to time this, but I wonder if I could have messed something up? The change landed in mono-2019-12, so I can retest when we are on it.

> in .NET 5 "everybody" will be using the same assemblies

This came up here: https://github.com/mono/mono/pull/18040#issuecomment-561780099

My thought was a build-time error would be better than a runtime error (and only on unhandled exception!). It could slightly help us migrate to .NET 5.