@jonathanpeppers 2020-03-10
Closing in favor of: https://github.com/xamarin/xamarin-android/pull/4302